Extract & expose api client from the process collector
