+++ pkg/process/checks/pod.go
@@ -53,5 +53,5 @@ func (c *PodCheck) Run(cfg *config.AgentConfig, groupID int32) ([]model.MessageB
 		return nil, err
 	}
 
-	return orchestrator.ProcessPodlist(podList, groupID, cfg)
+	return orchestrator.ProcessPodlist(podList, groupID, cfg, cfg.HostName, cfg.KubeClusterName)
 }
